pulse-robot-template-87375-main/fullstack/backend/requirements.txt
torch>=1.9.0
ultralytics>=8.0.0
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.21.0
